fix(tangle-dapp): Fix requested changes
