+++ libs/webb-ui-components/src/components/Dropdown/index.ts
@@ -2,6 +2,7 @@ export * from './Dropdown';
 export * from './DropdownBasicButton';
 export * from './DropdownBody';
 export * from './DropdownButton';
+export * from './DropdownField';
 export { default as DropdownMenu } from './DropdownMenu';
 export { default as DropdownMenuItem } from './DropdownMenuItem';
