--- libs/webb-ui-components/src/components/DropdownField/index.ts
@@ -1,1 +1,0 @@
-export { default as DropdownField } from './DropdownField';
