--- libs/webb-ui-components/src/components/DropdownField/types.ts
@@ -1,8 +1,0 @@
-export type DropdownFieldProps = {
-  title: string;
-  items: string[];
-  className?: string;
-  selectedItem: string;
-  setSelectedItem: (selectedItem: string) => void;
-  dropdownBodyClassName?: string;
-};
