+++ libs/webb-ui-components/src/components/index.ts
@@ -27,7 +27,6 @@ export * from './Disclaimer';
 export * from './Divider';
 export * from './Drawer';
 export * from './Dropdown';
-export * from './DropdownField';
 export * from './ErrorFallback';
 export { default as ExternalLinkIcon } from './ExternalLinkIcon';
 export { default as FeeDetails } from './FeeDetails';
